Extract minimal agent for reuse
